Musterlösung revisited

+++ Tag4/bmi/people_data.py
@@ -5,7 +5,7 @@ def read_person_data(path):
         people = list()
         for row in cleaned_rows:
             person_data = row.split(',')
-            person_dict = {'name': person_data[0], 'weight': int(person_data[1]), 'height': float(person_data[2])}
+            person_dict = {'name': person_data[0], 'weight': float(person_data[1]), 'height': float(person_data[2])}
             people.append(person_dict)
         return people
 def write_person_data(path, result_list):
